+++ .gitignore
@@ -36,4 +36,5 @@ proguard/
 .idea/workspace.xml
 .idea/tasks.xml
 .idea/libraries/
+.idea/codeStyleSettings.xml
 .DS_Store
