+++ app/build.gradle
@@ -1,4 +1,5 @@
 apply plugin: 'com.android.application'
+apply plugin: 'com.neenbedankt.android-apt'
 apply plugin: 'me.tatarka.retrolambda'
 
 android {
@@ -7,7 +8,7 @@ android {
 
     defaultConfig {
         applicationId 'mikazuki.android.app.feelingmatch'
-        minSdkVersion 14
+        minSdkVersion 19
         targetSdkVersion 23
         versionCode 1
         versionName '1.0'
@@ -25,6 +26,10 @@ android {
     }
 }
 
+repositories {
+    mavenCentral()
+}
+
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     testCompile 'junit:junit:4.12'
@@ -34,4 +39,9 @@ dependencies {
     compile 'com.android.support:animated-vector-drawable:23.2.1'
 
     compile 'com.jakewharton:butterknife:7.0.1'
+
+    compile 'org.parceler:parceler-api:1.0.4'
+    apt 'org.parceler:parceler:1.0.4'
+
+    compile 'com.annimon:stream:1.0.8'
 }
