+++ app/src/main/java/mikazuki/android/app/feelingmatch/model/User.java
@@ -2,13 +2,20 @@ package mikazuki.android.app.feelingmatch.model;
 
 import junit.framework.Assert;
 
+import org.parceler.Parcel;
+import org.parceler.Transient;
+
 /**
  * @author haijimakazuki
  */
+@Parcel
 public class User {
 
     private String name;
     private int sex; // girl=0, boy=1
+
+    public User() {
+    }
 
     public User(String name, int sex) {
         this.name = name;
@@ -33,10 +40,12 @@ public class User {
         this.sex = sex;
     }
 
+    @Transient
     public boolean isBoy() {
         return this.sex == 1;
     }
 
+    @Transient
     public boolean isGirl() {
         return this.sex == 0;
     }
