+++ app/src/main/java/mikazuki/android/app/feelingmatch/view/activity/BaseActivity.java
@@ -1,4 +1,4 @@
-package mikazuki.android.app.feelingmatch;
+package mikazuki.android.app.feelingmatch.view.activity;
 
 import android.support.annotation.IdRes;
 import android.support.v4.app.Fragment;
@@ -6,6 +6,8 @@ import android.support.v4.app.FragmentTransaction;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.AppCompatActivity;
+
+import mikazuki.android.app.feelingmatch.R;
 
 public abstract class BaseActivity extends AppCompatActivity {
