+++ app/src/main/res/menu/main_toolbar.xml
@@ -1,0 +1,9 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".view.activity.MainActivity">
+    <item
+        android:id="@+id/menu_start"
+        android:title="@string/start"
+        app:showAsAction="always" />
+</menu>
