+++ app/src/main/res/values/strings.xml
@@ -4,6 +4,6 @@
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
 
-    <string name="action_settings">Settings</string>
     <string name="remove_user">ユーザーの削除</string>
+    <string name="start">開始</string>
 </resources>
